Update jquery.dmst.js
Added option funcLoadOptions to allow a function to be used to load the options rather than a static associative array. Takes one parameter, a reference to the dmst instance, for loading directly to baseSelectOptions. Could probably use some iterations to find a more streamlined way of doing this...

Various minor updates.